Got metadata working mostly now. Pry needs a bit of massaging but will work.

+++ KodiLcdServer.h
@@ -58,7 +58,7 @@ public slots:
 signals:
     void connectionError();
     void apiError();
-    void progressPercentComplete(int);
+    void progressPercentComplete(double);
     void metaDataEnded();
 	void videoFormat(QString);
 	void audioFormat(QString);
